Rename utils/default.js -> utils/params.js
